+++ src/test/java/com/system/coin/exchange/service/CoinExchangeServiceTest.java
@@ -3,6 +3,7 @@ package com.system.coin.exchange.service;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -31,23 +32,23 @@ public class CoinExchangeServiceTest {
 	public void test_exchangeBills_success() throws Exception {
 		CoinExchangeRequest request = new CoinExchangeRequest();
 		request.setBill(1);
-		CoinCount coinCount = new CoinCount(0.1, 5);
+		CoinCount coinCount = new CoinCount(BigDecimal.valueOf(0.1), 5);
 		List<CoinCount> cointCountList = new ArrayList<CoinCount>();
 		cointCountList.add(coinCount);
 		request.setCoinCountList(cointCountList);
 		Mockito.doNothing().when(validatorService).validate(request);
-    	Map<Double, Integer> responseMap = coinExchangeService.exchange(request);
+    	Map<BigDecimal, Integer> responseMap = coinExchangeService.exchange(request);
     	assertNotNull(responseMap);  
-    	assertNotNull(responseMap.get(0.25));
-    	assertEquals(Integer.valueOf(2), responseMap.get(0.25));
-    	assertEquals(Integer.valueOf(5), responseMap.get(0.1));
+    	assertNotNull(responseMap.get(BigDecimal.valueOf(0.25)));
+    	assertEquals(Integer.valueOf(2), responseMap.get(BigDecimal.valueOf(0.25)));
+    	assertEquals(Integer.valueOf(5), responseMap.get(BigDecimal.valueOf(0.1)));
     }
     
     @Test
 	public void test_exchangeBills_failure() throws Exception {
 		CoinExchangeRequest request = new CoinExchangeRequest();
 		request.setBill(50);
-		CoinCount coinCount = new CoinCount(0.1, 5);
+		CoinCount coinCount = new CoinCount(BigDecimal.valueOf(0.1), 5);
 		List<CoinCount> cointCountList = new ArrayList<CoinCount>();
 		cointCountList.add(coinCount);
 		request.setCoinCountList(cointCountList);
